Some Changes

+++ lock.sh
@@ -4,8 +4,10 @@ read -s password
 for file in *
 #for file in *.{jpg,gif,png,jpeg}
 do
-echo $file
-   openssl enc -aes-256-cbc -salt -in $file -out "$file=enc" -k $password
-   rm -f $file    
+   if ! [[ $file == *.sh ]]; then
+	   echo $file
+	   openssl enc -aes-256-cbc -salt -in $file -out "$file=enc" -k $password
+	   rm -f $file  
+   fi  
 done;
